refactor: xddir json_serialized
